Refactor SEO codec
